media: vio: fix [XJ3-2874] Fix compile warning that may be a risk  no_3rd_skip
Summary: Fix compile warning that may be a risk

Test Plan: no

Reviewers: shengqiang.gao, mingang.lan, yongyong.duan, fenfen.wang, wenbo.he, yunqian.wang

Reviewed By: shengqiang.gao, mingang.lan, yongyong.duan, fenfen.wang, wenbo.he, yunqian.wang

Differential Revision: https://cr.hobot.cc/D163895

+++ drivers/media/platform/hobot/isp/v4l2_dev/src/fw_lib/acamera_command_api_impl.c
@@ -3910,7 +3910,7 @@ uint8_t sharpening_strength( acamera_fsm_mgr_t *instance, uint32_t value, uint8_
 #ifdef FR_OUTPUT_MODE_ID
 uint8_t fr_output_mode( acamera_fsm_mgr_t *instance, uint32_t value, uint8_t direction, uint32_t *ret_value )
 {
-    uint32_t fr_output_fmt = 0;
+    uint8_t fr_output_fmt = 0;
     *ret_value = 0;
     if ( direction == COMMAND_SET ) {
         switch ( value ) {
